[RLlib] Add StepFailedRecreateEnv exception (#55146)
## Why are these changes needed?

There exist environments that are inherently unstable (for example
because they rely on a system that is not under the control of the
user).

This PR introduces a special exception that a user-defined environment
can raise in case the external environment fails and requires a reset.
RLlib catches the exception and recreates the environment that raised
said exception without logging an error.